pyproject.toml
[project]
name = "task-daemon"
version = "0.1.0"
description = "A configurable task processing system with monitoring"
authors = [
    {name = "Jonathan James", email = "jonathanjamesm66@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "requests>=2.32.5",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.4",
    "msgpack>=1.1.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=5.26.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "black>=25.9.0",
    "flake8>=7.3.0",
]

[project.scripts]
task-daemon = "task_daemon.__main__:main"
generate-proto = "generate_proto:main"

[project.urls]
Homepage = "https://github.com/jona62/task-daemon"
Repository = "https://github.com/jona62/task-daemon"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_daemon"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "benchmark: marks tests as benchmark tests (deselect with '-m \"not benchmark\"')",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
